% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_input_email}
\alias{add_input_email}
\title{Specify an email address to which input feedback can be sent}
\usage{
add_input_email(disp, email)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}

\item{email}{An email address.}
}
\description{
Specify an email address to which input feedback can be sent
}
