% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_input_vars}
\alias{add_input_vars}
\title{Specify meta variables whose values should be provided in input feedback}
\usage{
add_input_vars(disp, vars)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}

\item{vars}{A vector of meta variable names found in the display.}
}
\description{
Specify meta variables whose values should be provided in input feedback
}
