% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_meta_def}
\alias{add_meta_def}
\title{Add a meta variable definition to a trelliscope display}
\usage{
add_meta_def(trdf, obj)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}} or a
data frame which will be cast as such.}

\item{obj}{A meta variable definition created with a meta_*() function.}
}
\description{
Add a meta variable definition to a trelliscope display
}
