% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_meta_defs}
\alias{add_meta_defs}
\title{Add multiple meta variable definitions to a trelliscope display}
\usage{
add_meta_defs(trdf, ...)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope]{as_trelliscope()}} or a
data frame which will be cast as such.}

\item{...}{Any number of objects created with meta_*() functions.}
}
\description{
Add multiple meta variable definitions to a trelliscope display
}
