% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_meta_tags}
\alias{add_meta_tags}
\title{Specify tags for meta variables}
\usage{
add_meta_tags(disp, ...)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}

\item{...}{A named set of vectors of tags, where each name must correspond to
one of the variables in the dataset}
}
\description{
Specify tags for meta variables
}
\details{
This function can be useful if you don't want to go to the trouble
of explicitly setting meta variable definitions but still want variable
tags.
}
