% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{add_trelliscope_resource_path}
\alias{add_trelliscope_resource_path}
\title{Add Trelliscope resource path for Shiny app}
\usage{
add_trelliscope_resource_path(prefix, path)
}
\arguments{
\item{prefix, path}{See \code{\link[shiny:resourcePaths]{shiny::addResourcePath()}}}
}
\description{
Add Trelliscope resource path for Shiny app
}
