% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trelliscope-data.R
\docType{data}
\name{airplane}
\alias{airplane}
\title{Airplane Data}
\format{A data frame with 1479 rows and 47 variables}
\usage{
data(airplane)
}
\description{
Airplane data
}
\details{
\itemize{
  \item dat_ams.      : POSIXct, format: "2010-06-28 16:25:21" "2010-06-28 16:25:34" ...
  \item org.          : num  9.26 13.29 12.13 12.36 10.11 ...
  \item NO3.          : num  0.449 0.589 0.536 0.57 0.457 ...
  \item SO4.          : num  0.115 0.193 0.232 0.307 0.197 ...
  \item NH4.          : num  0.126 0.229 0.224 0.179 0.172 ...
  \item CHL.          : num  0.01119 0.01748 0.00607 0.03346 0.00683 ...
  \item org_err.      : num  0.111 0.117 0.116 0.116 0.115 ...
  \item NO3_err.      : num  0.0172 0.0181 0.0177 0.0182 0.0175 ...
  \item SO4_err.      : num  0.0222 0.0244 0.0252 0.0253 0.025 ...
  \item NH4_err.      : num  0.0331 0.0331 0.0331 0.0332 0.0331 ...
  \item CHl_err.      : num  0.0124 0.0122 0.0123 0.0123 0.0123 ...
  \item latitude.     : num  38.7 38.7 38.7 38.7 38.7 ...
  \item longitude.    : num  -121 -121 -121 -121 -121 ...
  \item altitude.     : num  354 369 385 393 388 ...
  \item static_p.     : num  968 966 964 963 964 ...
  \item altitude_p.   : num  327 341 356 365 360 ...
  \item static_t.     : num  28.4 28.7 28.6 28.5 28.8 ...
  \item potential_t.  : num  31.3 31.7 31.7 31.7 32 ...
  \item dewpt.        : num  14.1 13.5 13.4 12.9 11.1 ...
  \item RH.           : num  41.6 39.3 39.3 38.5 33.5 ...
  \item h2o_MF.       : num  10.54 10.11 10.05 9.8 8.68 ...
  \item windspd.      : num  3.45 3.9 4.14 4.12 4.7 4.54 3.42 3.1 2.98 2.64 ...
  \item winddir.      : num  330 325 320 318 317 ...
  \item cpc_3025.     : num  5508 3915 4037 3969 3567 ...
  \item cpc_3010.     : num  483 352 364 348 314 ...
  \item ozone.        : num  46.8 44.3 45.2 45.5 40.4 ...
  \item co.           : num  109.7 97.6 101.2 103.1 91.8 ...
  \item so2.          : num  NA NA NA NA NA NA NA 0.6 0.465 0.644 ...
  \item no.           : num  0.3074 0.0514 0.1456 0.1593 0.115 ...
  \item no2.          : num  1.745 0.358 0.573 0.688 0.8 ...
  \item noy.          : num  5.25 3.48 4.09 4.02 3.53 ...
  \item mz_33.        : num  16.7 14.6 14.4 14.1 11.1 ...
  \item mz_42.        : num  0.1024 0.1176 0.1333 0.0916 0.0407 ...
  \item mz_45.        : num  5.98 5.22 5.19 4.84 5.07 ...
  \item mz_59.        : num  8.49 7.71 6.85 7.08 6.46 ...
  \item mz_69.        : num  1.589 0.811 0.791 0.904 0.911 ...
  \item mz_71.        : num  3.88 3.65 3.35 2.94 2.74 ...
  \item mz_79.        : num  0.417 0.243 0.192 0.261 0.178 ...
  \item mz_93.        : num  0.446 0.571 0.454 0.218 0.185 ...
  \item mz_107.       : num  0.686 0.869 0.661 0.78 0.265 ...
  \item mz_137.       : num  0.519 0.452 0.093 0.353 0.348 ...
  \item mz_43.        : num  8.52 8.3 8.17 8.38 6.87 ...
  \item mz_61.        : num  5.39 5.13 5.1 5.17 4.96 ...
  \item mz_87.        : num  0.302 0.235 0.418 0.298 0.364 ...
  \item mz69_71.      : num  5.47 4.46 4.14 3.84 3.65 ...
  \item leg_flag.     : int  13 13 13 13 3 3 3 3 3 3 ...
  \item nox.          : num  2.053 0.41 0.718 0.847 0.915 ...
}
}
\keyword{datasets}

