% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_cog_methods.R
\name{cogNcol}
\alias{cogNames}
\alias{cogNcol}
\alias{cogNrow}
\alias{getCogData}
\title{Methods for Cognostics Connections}
\usage{
cogNcol(x, ...)

cogNrow(x, ...)

cogNames(x, ...)

getCogData(x, rowIdx, colIdx, ...)
}
\arguments{
\item{x}{cognostics connection object}

\item{...}{other objects passed onto generic methods}

\item{rowIdx}{index of rows to be retrieved from the cognostics connection}

\item{colIdx}{index of columns to be retrieved from the cognostics connection}
}
\description{
Methods for Cognostics Connections
}
\note{
These methods are used mainly by the trelliscope viewer and therefore must be exported.  Their purpose is to provide a general interface for a cognostics store.  Currently just data frames are used for cognostics, but in previous versions systems like MongoDB were used.  These methods should never need to be used by an analyst.
}
\examples{
# see examples for makeDisplay()
}
\seealso{
\code{\link{makeDisplay}}
}

