% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{copyRFromPackage}
\alias{copyRFromPackage}
\title{Copy R Files From Package}
\usage{
copyRFromPackage(from_package, conn = getOption("vdbConn"))
}
\arguments{
\item{from_package}{}

\item{conn}{vdbConn}
}
\description{
Just a helper function that wraps copyRSource
}

