% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{copyRSource}
\alias{copyRSource}
\title{Copy files ending in .R from source directory}
\usage{
copyRSource(from_dir, conn = getOption("vdbConn"))
}
\arguments{
\item{from_dir}{directory with R source}

\item{conn}{vdb connection}
}
\description{
These R files will be sourced in the global environment of the trelliscope viewer.
Useful for getting custom functions into the global environment.
}
\examples{
copyRSource(".") #copies R files from current directory

}
\author{
Jeremiah Rounds
}

