% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webSync.R
\name{deployToShinyApps}
\alias{deployToShinyApps}
\title{Deploy VDB to shinyapps.io}
\usage{
deployToShinyApps(vdbConn = getOption("vdbConn"), appName = NULL,
  account = NULL, redeploy = TRUE, size = NULL, instances = NULL,
  quiet = FALSE)
}
\arguments{
\item{vdbConn}{A vdbConn object containing the VDB connection settings}

\item{appName}{name of application (app will be available at https://[account].shinyapps.io/[appName]/) - if not supplied, will use the name of VDB connection}

\item{account}{passed to \code{shinyapps::configureApp}}

\item{redeploy}{passed to \code{shinyapps::configureApp}}

\item{size}{passed to \code{shinyapps::configureApp}}

\item{instances}{passed to \code{shinyapps::configureApp}}

\item{quiet}{passed to \code{shinyapps::configureApp}}
}
\description{
Deploy VDB to shinyapps.io
}
\details{
If you do not have a shinyapps.io account and have not set your account info, first visit here prior to calling this function: \url{http://shiny.rstudio.com/articles/shinyapps.html}.
}
\author{
Ryan Hafen

\code{\link{syncLocalData}}
}

