% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{elnino}
\alias{elnino}
\title{El Nino Title TODO}
\format{A data frame with 178327 rows and 9 variables}
\usage{
data(elnino)
}
\description{
El Nino details TODO
}
\details{
\itemize{
  \item buoy.     : chr  "buoyid_01" "buoyid_01" "buoyid_01" "buoyid_01" ...
  \item date.     : Date, format: "1992-01-01" "1992-09-22" ...
  \item lat.      : num  NA -7.97 -7.97 -7.97 -7.97 -7.97 -7.97 -7.97 -7.97 -7.97 ...
  \item lon.      : num  NA 55 55 55 55 ...
  \item zonWinds. : num  NA -4.6 -4.5 -5.4 -4.8 -5.9 -7 -6.6 -6.6 -7.6 ...
  \item merWinds. : num  NA -0.2 -2.1 -1.3 -0.6 -0.3 -0.6 0.1 1.8 0.2 ...
  \item humidity. : num  NA NA NA NA NA NA NA NA NA NA ...
  \item airTemp.  : num  NA NA NA NA NA NA NA NA NA NA ...
  \item ssTemp.   : num  NA NA 26.4 26.4 26.3 ...
}
}
\keyword{datasets}

