% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{elninoMap}
\alias{elninoMap}
\title{El Nino Map Title TODO}
\format{A list of class 'map' of 4 variables}
\usage{
data(elninoMap)
}
\description{
El Nino Map details TODO
}
\details{
\itemize{
  \item x    : num [1:27121] 50 49.8 48.1 48 47.7 ...
  \item y    : num [1:27121] 55.9 56.1 56.7 57 57.2 ...
  \item range: num [1:4] -180 190.3 -85.4 83.6
  \item names: chr [1:2284] "Canada" "South Africa" "Denmark" "Great Lakes:Superior, Huron, Michigan" ...
}
}
\keyword{datasets}

