% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDisplay_internals.R
\name{encodePNG}
\alias{encodePNG}
\title{base64 Encoding of a .png File}
\usage{
encodePNG(plotLoc)
}
\arguments{
\item{plotLoc}{location of a png file on disk to encode as a base64 string}
}
\description{
base64 Encoding of a .png File
}
\examples{
f <- tempfile(fileext = ".png")
png(f)
plot(1:10)
dev.off()
encodePNG(f)
}

