% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_trelliscope.R
\name{facet_trelliscope}
\alias{facet_trelliscope}
\title{Add a trelliscope facet to a ggplot}
\usage{
facet_trelliscope(
  facets,
  scales = "same",
  add_plot_metrics = FALSE,
  data = ggplot2::waiver()
)
}
\arguments{
\item{facets}{A formula to facet the panels on. Similar to
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}'s `facets``}

\item{scales}{Should scales be the same (\code{"same"}, the default),
free (\code{"free"}), or sliced (\code{"sliced"}). May provide a single string or
two strings, one for the X and Y axis respectively.}

\item{add_plot_metrics}{Should metrics about each panel be automatically
calculated? These metrics are based on the context of what is being
plotted, e.g. correlation coefficient if plot is a scatterplot.}

\item{data}{data used for faceting. Defaults to the first layer data}
}
\description{
Add a trelliscope facet to a ggplot
}
