% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_variables.R
\name{filter_cat_href}
\alias{filter_cat_href}
\title{Create a link to another display filtered to categories of a variable}
\usage{
filter_cat_href(
  display_name,
  varname,
  values = NULL,
  labels = NULL,
  prefix = NULL
)
}
\arguments{
\item{display_name}{The name of the display to link to.}

\item{varname}{The name of the variable to filter on.}

\item{values}{The values of the variable to filter on.}

\item{labels}{Labels in the linked display to show.}

\item{prefix}{A prefix to add to the link. Not needed if the display is part
of the same app.}
}
\description{
Create a link to another display filtered to categories of a variable
}
