% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_wrappers.R
\name{filter_range}
\alias{filter_range}
\title{Specify a range "filter" state (applies to numeric, date, and datetime
meta variables)}
\usage{
filter_range(varname, min = NULL, max = NULL)
}
\arguments{
\item{varname}{The name of the variable to filter on.}

\item{min}{Lower bound of the range (if NULL, there is no lower bound).}

\item{max}{Upper bound of the range (if NULL, there is no upper bound).}
}
\description{
Specify a range "filter" state (applies to numeric, date, and datetime
meta variables)
}
