% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_wrappers.R
\name{filter_string}
\alias{filter_string}
\title{Specify a string "filter" state (applies to string and factor meta
variables)}
\usage{
filter_string(varname, regexp = NULL, values = NULL)
}
\arguments{
\item{varname}{The name of the variable to filter on.}

\item{regexp}{A search string that can be a regular expression indicating
the values of the variable to filter on.}

\item{values}{A vector of specific values of the variable to filter on. If
\code{values} is specified, \code{regexp} will be ignored.}
}
\description{
Specify a string "filter" state (applies to string and factor meta
variables)
}
