% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gap}
\alias{gap}
\title{Gapminder data with ISO codes and country centroids}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1704 rows and 9 columns.
}
\source{
\url{https://cran.r-project.org/web/packages/gapminder/index.html}
}
\usage{
gap
}
\description{
Gapminder data with ISO codes and country centroids
}
\keyword{datasets}
