% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gapminder data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1476 rows and 9 columns.
}
\source{
\url{https://cran.r-project.org/web/packages/gapminder/index.html}
}
\usage{
gapminder
}
\description{
Gapminder data
}
\keyword{datasets}
