% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_panels.R
\name{get_test_subset}
\alias{get_test_subset}
\title{Get a subset of a dataset to test a plot function on}
\usage{
get_test_subset(full_dat, summ_dat, by = NULL)
}
\arguments{
\item{full_dat}{The full dataset from which to extract a subset}

\item{summ_dat}{The summary dataset to which a plot column will be added.
Each row of this dataset will correspond to one subset of \code{full_dat} and a
sample row will be used to produce a single test subset.}

\item{by}{A list of variables found in both \code{full_dat} and in \code{summ_dat}
to which a plot column will be added. This is used to specify which subset
of \code{data} to apply for a given plot. If not provided, it is inferred based
on the variables found in both \code{data} and the summary data.}
}
\description{
Get a subset of a dataset to test a plot function on
}
