% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{globalsExist}
\alias{globalsExist}
\title{Check to see if the global data file already exist}
\usage{
globalsExist(conn = getOption("vdbConn"))
}
\arguments{
\item{conn}{vdb connection}
}
\value{
TRUE/FALSE
}
\description{
Does not check to see if any particular data set is in the global file.
}
\examples{
globalsExist()
}

