% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{globalsFile}
\alias{globalsFile}
\title{Path to global rdata in Trelliscope viewer.}
\usage{
globalsFile(conn = getOption("vdbConn"))
}
\arguments{
\item{conn}{vdb connection}
}
\value{
character file name
}
\description{
\code{globalsFile} returns an appropriate file name to save to.
Objects in this rdata will be loaded by the server viewer code prior to showing 
displays
}
\examples{
save(foo, file=globalsPath(conn=conn))
}

