% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_types.R
\name{iframe_panel}
\alias{iframe_panel}
\title{Cast a vector of URLs pointing to html pages as an image panel source}
\usage{
iframe_panel(x)
}
\arguments{
\item{x}{a vector of URLs pointing to html pages}
}
\description{
Cast a vector of URLs pointing to html pages as an image panel source
}
