% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_types.R
\name{iframe_panel_local}
\alias{iframe_panel_local}
\title{Cast a vector of URLs pointing to html pages as an image panel source}
\usage{
iframe_panel_local(x)
}
\arguments{
\item{x}{a vector of URLs pointing to html pages}
}
\description{
Cast a vector of URLs pointing to html pages as an image panel source
}
\note{
\code{x} must be paths relative to the \code{path} argument passed to
\code{\link{trelliscope}}.
}
