% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_types.R
\name{img_panel}
\alias{img_panel}
\title{Cast a vector of URLs pointing to images as an image panel source}
\usage{
img_panel(x, aspect_ratio = 1.5)
}
\arguments{
\item{x}{A vector of URLs pointing to images.}

\item{aspect_ratio}{Specifies the aspect ratio (width / height) to use when
displaying the image in the browser.}
}
\description{
Cast a vector of URLs pointing to images as an image panel source
}
