% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer_meta}
\alias{infer_meta}
\title{Infer meta variable definitions}
\usage{
infer_meta(disp)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}
}
\description{
Infer meta variable definitions
}
