% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_checkbox}
\alias{input_checkbox}
\title{Specify a "checkbox" input}
\usage{
input_checkbox(name, label = name, active = TRUE, options)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{options}{A vector of checkbox options.}
}
\description{
Specify a "checkbox" input
}
\examples{
library(ggplot2)
library(dplyr)

panel_dat <- (
  ggplot(gapminder, aes(year, lifeExp)) +
    geom_point() +
    facet_panels(vars(country, continent))
  ) |>
  as_panels_df()

trell <- panel_dat |>
as_trelliscope_df() |>
add_inputs(
  input_checkbox(
    name = "Checkbox Input",
    label = "A space to add custom button inputs",
    options = c("yes", "no")
  ),
  email = "johndoe@email.com"
)
\dontrun{
view_trelliscope(trell)
}
}
\seealso{
Other {input types}: 
\code{\link{input_multiselect}()},
\code{\link{input_number}()},
\code{\link{input_radio}()},
\code{\link{input_select}()},
\code{\link{input_text}()}
}
\concept{{input types}}
