% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_select}
\alias{input_select}
\title{Specify a "select" input}
\usage{
input_select(name, label = name, active = TRUE, options)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{options}{A vector of options for the select dropdown.}
}
\description{
Specify a "select" input
}
\examples{
library(ggplot2)
library(dplyr)

panel_dat <- (
  ggplot(gap, aes(year, life_exp)) +
    geom_point() +
    facet_panels(vars(country, continent))
  ) |>
  as_panels_df()

trell <- panel_dat |>
as_trelliscope_df() |>
add_inputs(
  input_select(
    name = "Select Input",
    label = "A space to add custom dropdown inputs",
    options =c("yes", "no")
  )
)
\dontrun{
view_trelliscope(trell)
}
}
\seealso{
Other {input types}: 
\code{\link{input_checkbox}()},
\code{\link{input_multiselect}()},
\code{\link{input_number}()},
\code{\link{input_radio}()},
\code{\link{input_text}()}
}
\concept{{input types}}
