% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_text}
\alias{input_text}
\title{Specify a "text box" input}
\usage{
input_text(name, label = name, active = TRUE, width = 80, height = 3)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{width}{Width (in characters) of the text box input.}

\item{height}{Height (in lines of text) of the text box input.}
}
\description{
Specify a "text box" input
}
\seealso{
Other {input types}: 
\code{\link{input_checkbox}()},
\code{\link{input_multiselect}()},
\code{\link{input_number}()},
\code{\link{input_radio}()},
\code{\link{input_select}()}
}
\concept{{input types}}
