% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{local_websocket_server}
\alias{local_websocket_server}
\title{Set up a local websocket server}
\usage{
local_websocket_server(
  d,
  plot_fn,
  width = 500,
  height = 500,
  format = "png",
  force = FALSE
)
}
\arguments{
\item{d}{A data frame from which subsets will be extracted and plots will
be made. Should be a superset of the summary dataset being cast as a
trelliscope data frame.}

\item{plot_fn}{A function that produces a panel from a given subset of \code{d}}

\item{width}{Width in pixels of each panel.}

\item{height}{Height in pixels of each panel.}

\item{format}{The format of the panels the server will provide. Can be
one of "png" , "svg", or "html".}

\item{force}{Should server force panels to be written? If \code{FALSE}, if the
panel has already been generated, that is what will be made available.}
}
\description{
Set up a local websocket server
}
