% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{map_plot}
\alias{map_plot}
\title{Apply a function to each element of a vector and return a vector of plots}
\usage{
map_plot(.x, .f, ...)
}
\arguments{
\item{.x}{a list or atomic vector (see \code{\link[purrr:map]{purrr::map()}} for details)}

\item{.f}{a function, formula, or atomic vector (see \code{\link[purrr:map]{purrr::map()}} for
details)}

\item{...}{additional arguments passed on to .f (see \code{\link[purrr:map]{purrr::map()}} for
details)}
}
\description{
Apply a function to each element of a vector and return a vector of plots
}
\details{
See \code{\link[purrr:map]{purrr::map()}}
}
\examples{
\donttest{
library(dplyr)
library(tidyr)
library(purrr)

x <- ggplot2::mpg |>
  tidyr::nest(data = !dplyr::one_of(c("manufacturer", "class"))) |>
  dplyr::mutate(panel = map_plot(data, function(x) {
    ggplot2::qplot(hwy, cty, data = x)
  })) |>
  trelliscope(name = "test")
}
}
