% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mars_rover}
\alias{mars_rover}
\title{Mars rover data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1245 rows and 5 columns.
}
\source{
\url{https://github.com/corincerami/mars-photo-api}
}
\usage{
mars_rover
}
\description{
Mars rover data
}
\keyword{datasets}
