% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_currency}
\alias{meta_currency}
\title{Specify a "currency" metadata variable}
\usage{
meta_currency(
  varname,
  label = NULL,
  tags = NULL,
  code = "USD",
  digits = 2,
  log = NULL
)
}
\arguments{
\item{varname}{Name of the variable.}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{code}{Currency code. See \link{currencies} for a list of possibilities.}

\item{digits}{How many digits to round to when displaying the number.
If \code{NULL}, all digits will be shown. Default is 2.}

\item{log}{Should the variable's distribution be shown on the log scale?
If not specified, an inference will be made based on its values.}
}
\description{
Specify a "currency" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_geo}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
