% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_datetime}
\alias{meta_datetime}
\title{Specify a "datetime" metadata variable}
\usage{
meta_datetime(varname, description = varname, tags = NULL)
}
\arguments{
\item{varname}{Name of the variable.}

\item{description}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}
}
\description{
Specify a "datetime" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_geo}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
