% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_factor}
\alias{meta_factor}
\title{Specify a "factor" metadata variable}
\usage{
meta_factor(varname, label = NULL, tags = NULL, levels = NULL)
}
\arguments{
\item{varname}{Name of the variable.}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{levels}{A vector of factor levels the correspond to the variable.
The order of levels will be respected when sorted, etc. If NULL,
levels will be inferred from the data based on the order they appear.}
}
\description{
Specify a "factor" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_geo}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
