% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_geo}
\alias{meta_geo}
\title{Specify a "geo" metadata variable}
\usage{
meta_geo(varname, label = NULL, tags = NULL, latvar, longvar)
}
\arguments{
\item{varname}{Name of the new variable that will act as geographical
coordinates (cannot exist in dataset).}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{latvar}{Name of variable that contains the latitude.}

\item{longvar}{Name of variable that contains the longitude.}
}
\description{
Specify a "geo" metadata variable
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

panel_dat <- (ggplot(gapminder, aes(year, lifeExp)) +
                geom_point() +
                facet_panels(~country + continent)) |>
  nest_panels() 


meta_dat <- gapminder|>
  group_by(country, continent) |>
  summarise(
    mean_lifeexp = mean(lifeExp),
    min_lifeexp = min(lifeExp),
    max_lifeexp = max(lifeExp),
    mean_gdp = mean(gdpPercap),
    first_year = min(year),
    latitude = first(latitude),
    longitude = first(longitude),
    .groups = "drop"
  ) |>
  ungroup() |>
  mutate(
    first_date = as.Date(paste0(first_year, "-01-01")),
    first_datetime = as.POSIXct.Date(first_date),
    continent = as.factor(continent),
    wiki_link = paste0("https://en.wikipedia.org/wiki/", country)
  )

joined_dat <- left_join(panel_dat, meta_dat, 
                        by = join_by(country, continent))

trell <- joined_dat |>
  as_trelliscope_df() |>
  write_panels() |>
  meta_geo("coordinates",
           label = "Lat and Long of the country's capital",
           tags = "Location",
           latvar = "latitude",
           longvar = "longitude")) |>
write_trelliscope() |>
view_trelliscope()
}


}
\seealso{
Other {metadata types}: 
\code{\link{meta_currency}()},
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
