% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_graph}
\alias{meta_graph}
\title{Specify a "graph" metadata variable}
\usage{
meta_graph(
  varname,
  label = NULL,
  tags = NULL,
  idvarname,
  linkidvarname,
  labelvarname = idvarname,
  direction = c("none", "to", "from")
)
}
\arguments{
\item{varname}{Name of the new variable that will act as a graph
(cannot exist in dataset).}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{idvarname}{Name of the variable in the data that identifies the
node that each entity belongs to.}

\item{linkidvarname}{Name of the variable in the data that contains the
identifier of the node that each entity links to}

\item{labelvarname}{Name of the variable in the data that is used to
label the nodes in the graph.}

\item{direction}{Direction of the links specifed in \code{varname}. One of
"none", "to", or "from". Determines whether and how arrows are shown
in the network graph in the app.}
}
\description{
Specify a "graph" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_currency}()},
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_geo}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
