% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_number}
\alias{meta_number}
\title{Specify a "number" metadata variable}
\usage{
meta_number(
  varname,
  label = NULL,
  tags = NULL,
  digits = NULL,
  locale = TRUE,
  log = NULL
)
}
\arguments{
\item{varname}{Name of the variable.}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{digits}{How many digits to round to when displaying the number.
If not specified, a value will be inferred. If -1, all digits will be shown.}

\item{locale}{Should the variable be displayed using its locale?
For example, 1234.56 in US would be displayed as 1,234.56.}

\item{log}{Should the variable's distribution be shown on the log scale?
If not specified, an inference will be made based on its values.}
}
\description{
Specify a "number" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_currency}()},
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_geo}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
