% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_panel}
\alias{meta_panel}
\title{Specify a "string" metadata variable}
\usage{
meta_panel(
  varname,
  label = NULL,
  tags = NULL,
  type = NULL,
  aspect = NULL,
  source = "file"
)
}
\arguments{
\item{varname}{Name of the variable.}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{type}{The type of panel. One of "img" or "iframe".}

\item{aspect}{The aspect ratio at which the panel should be displayed.
Must be a positive number indicating width / height.}

\item{source}{The source for the panel. Must be one of "file", "websocket",
or "REST".}
}
\description{
Specify a "string" metadata variable
}
