% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_string}
\alias{meta_string}
\title{Specify a "string" metadata variable}
\usage{
meta_string(varname, description = varname, tags = "common")
}
\arguments{
\item{varname}{Name of the variable.}

\item{description}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}
}
\description{
Specify a "string" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_geo}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()}
}
\concept{{metadata types}}
