% Generated by roxygen2 (4.0.1): do not edit by hand
\name{myRunApp}
\alias{myRunApp}
\title{internal runApp (from shiny's runApp() - need to be able to add hash string)}
\usage{
myRunApp(appDir = getwd(), port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"), workerId = "",
  quiet = FALSE, display.mode = c("auto", "normal", "showcase"),
  hash = "")
}
\arguments{
\item{appDir,port,launch.browser,host,workerId,quiet,display.mode,hash}{TODO}
}
\description{
View a display or run Shiny display viewer
}
\author{
Ryan Hafen
}

