% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_variables.R
\name{number}
\alias{number}
\title{Create a number vector}
\usage{
number(x = double(), digits = 2, locale = TRUE, log = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{How many digits to round to when displaying the number.
If not specified, a value will be inferred. If -1, all digits will be shown.}

\item{locale}{Should the variable be displayed using its locale?
For example, 1234.56 in US would be displayed as 1,234.56.}

\item{log}{Should the variable's distribution be shown on the log scale?
If not specified, an inference will be made based on its values.}
}
\description{
Create a number vector
}
