% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_types.R
\name{panel_calc}
\alias{panel_calc}
\title{Add a plot column to a dataset}
\usage{
panel_calc(plot_fn, data, by = NULL, cur = dplyr::pick(dplyr::everything()))
}
\arguments{
\item{plot_fn}{A function that produces a panel from a given subset of \code{data}}

\item{data}{A data frame from which subsets will be extracted and plots will
be made. Should be a superset of the summary dataset to which this plot
column is being added.}

\item{by}{A list of variables found in both \code{data} and in the summary
dataset to which this plot column is being added. This is used to specify
which subset of \code{data} to apply for a given plot.}
}
\description{
Add a plot column to a dataset
}
