% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_panels.R
\name{panel_local}
\alias{panel_local}
\title{Add a "panel_local" column to a dataset}
\usage{
panel_local(x = character())
}
\arguments{
\item{x}{A character vector of paths to local files to be used as panels.}
}
\description{
Add a "panel_local" column to a dataset
}
