% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_trelliscope.R
\name{panel_options_lazy}
\alias{panel_options_lazy}
\title{Specify options for lazily-rendered panels in a Trelliscope display}
\usage{
panel_options_lazy(
  width = 500,
  height = 500,
  format = NULL,
  force = FALSE,
  prerender = TRUE
)
}
\arguments{
\item{width}{Width in pixels of each panel.}

\item{height}{Height in pixels of each panel.}

\item{format}{The format of the panels the server will provide. Can be
one of "png" , "svg", or "html".}

\item{force}{Should server force panels to be written? If \code{FALSE}, if the
panel has already been generated, that is what will be made available.}

\item{prerender}{If "TRUE", lazy panels will be rendered prior to viewing
the display. If "FALSE", a local R websockets server will be created and
plots will be rendered on the fly when requested by the app. The latter
is only available when using Trelliscope locally.}
}
\description{
Specify options for lazily-rendered panels in a Trelliscope display
}
\examples{
# TODO
}
