% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_write.R
\name{panels_write}
\alias{panels_write}
\title{Write panels}
\usage{
panels_write(disp, width = 500, height = 500, format = "png")
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}

\item{width}{Width in pixels of each panel.}

\item{height}{Height in pixels of each panel.}

\item{format}{The format of the image if it is not an htmlwidget. Can be
either "png" or "svg".}
}
\description{
Write panels
}
\note{
The size of panels will vary when shown in the viewer, but here the
specification of height and width help determine the plot aspect ratio
as well as the initial resolution to render plot text, etc. with.
}
