% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb.R
\name{phantomInstall}
\alias{phantomInstall}
\title{Get instructions on how to install phantomjs}
\usage{
phantomInstall()
}
\description{
Get instructions on how to install phantomjs
}
\examples{
phantomInstall()
}

