% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{plot_column}
\alias{plot_column}
\title{Add a plot column to a dataset}
\usage{
plot_column(
  plot_fn,
  data,
  by,
  width = 500,
  height = 500,
  format = "png",
  force = FALSE,
  prerender = TRUE
)
}
\arguments{
\item{plot_fn}{A function that produces a panel from a given subset of \code{data}}

\item{data}{A data frame from which subsets will be extracted and plots will
be made. Should be a superset of the summary dataset to which this plot
column is being added.}

\item{by}{A list of variables found in both \code{data} and in the summary
dataset to which this plot column is being added. This is used to specify
which subset of \code{data} to apply for a given plot.}

\item{width}{Width in pixels of each panel.}

\item{height}{Height in pixels of each panel.}

\item{format}{The format of the panels the server will provide. Can be
one of "png" , "svg", or "html".}

\item{force}{Should server force panels to be written? If \code{FALSE}, if the
panel has already been generated, that is what will be made available.}

\item{prerender}{If "TRUE", the plots must be rendered prior to viewing the
display. If "FALSE", a local R websockets server will be created and plots
will be rendered on the fly when requested by the app. The latter is only
available when using Trelliscope locally.}
}
\description{
Add a plot column to a dataset
}
