% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{print.cog}
\alias{print.cog}
\title{Print a cognostics object}
\usage{
\method{print}{cog}(x, ...)
}
\arguments{
\item{x}{a cognostics object}

\item{\ldots}{further arguments passed to or from other methods}
}
\description{
Print a cognostics object
}

