% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_cog_df.R
\name{print.dfCogConn}
\alias{print.dfCogConn}
\title{Print a dfCogConn object}
\usage{
\method{print}{dfCogConn}(x, ...)
}
\arguments{
\item{x}{a "dfCogConn" object}

\item{\ldots}{further arguments passed to or from other methods}
}
\description{
Print a dfCogConn object
}

