% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/displayObj.R
\name{print.displayObj}
\alias{print.displayObj}
\title{S3method print displayObj}
\usage{
\method{print}{displayObj}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
S3method print displayObj
}

