% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtrellis.R
\name{print.qtrellis}
\alias{print.qtrellis}
\title{Print a qtrellis Object}
\usage{
\method{print}{qtrellis}(x, ...)
}
\arguments{
\item{x}{a "qtrellis" object}

\item{\ldots}{further arguments passed to or from other methods}
}
\description{
Print a qtrellis Object
}

