% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_vdb.R
\name{print.vdbConn}
\alias{print.vdbConn}
\title{Print a vdbConn Object}
\usage{
\method{print}{vdbConn}(x, ...)
}
\arguments{
\item{x}{a "vdbConn" object}

\item{\ldots}{further arguments passed to or from other methods}
}
\description{
Print a vdbConn Object
}
\examples{
conn <- vdbConn(tempfile(), name = "myvdb", autoYes = TRUE)
conn
}

