% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rover_icon_b64}
\alias{rover_icon_b64}
\title{Base-64 encoded icon of a rover for use in the Mars rover example}
\format{
An object of class \code{character} of length 1.
}
\usage{
rover_icon_b64
}
\description{
Base-64 encoded icon of a rover for use in the Mars rover example
}
\keyword{datasets}
