% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{saveGlobals}
\alias{saveGlobals}
\title{Save global environment data for displays}
\usage{
saveGlobals(..., conn = getOption("vdbConn"))
}
\arguments{
\item{...}{objects to be saved in globals.Rdata for trelliscope displays.}

\item{conn}{vdb connection}
}
\description{
BUG: This function doesn't work as intended because of scope for ...
Makes data available in the global environment of all displays of a connection in trelliscope
}
\details{
Useful to avoid bloating individual displays with global variables common to all displays in a data base.
}
\examples{
data(iris)
saveGlobals(iris)
}
\author{
Jeremiah Rounds
}

