% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_default_filters}
\alias{set_default_filters}
\title{Add a filter state specifications to a trelliscope display}
\usage{
set_default_filters(trdf, ..., add = TRUE)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}} or a
data frame which will be cast as such.}

\item{...}{Filter state specifications (e.g. \code{\link[=filter_string]{filter_string()}},
\code{\link[=filter_range]{filter_range()}}).}

\item{add}{Should existing filter state specifications be added to?
Default is TRUE. If FALSE, the entire sort specification will be overridden.}
}
\description{
Add a filter state specifications to a trelliscope display
}
