% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_default_labels}
\alias{set_default_labels}
\title{Add a labels state specification to a trelliscope display}
\usage{
set_default_labels(trdf, varnames)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{varnames}{A vector of variable names whose values should appear
as labels. If NULL, no labels will be displayed.
when viewing the display.}
}
\description{
Add a labels state specification to a trelliscope display
}
