% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_default_sort}
\alias{set_default_sort}
\title{Add a labels state specification to a trelliscope display}
\usage{
set_default_sort(trdf, varnames, dirs = "asc", add = FALSE)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{varnames}{A vector of variable names to sort on.}

\item{dirs}{A vector of directions to sort on ("asc" or "desc").}

\item{add}{Should an existing sort specification be added to? If FALSE
(default), the entire sort specification will be overridden.}
}
\description{
Add a labels state specification to a trelliscope display
}
