% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_filters}
\alias{set_filters}
\title{Add a filter state specifications to a trelliscope display}
\usage{
set_filters(disp, ..., add = TRUE)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}

\item{...}{Filter state specifications (e.g. \code{\link[=filter_string]{filter_string()}},
\code{\link[=filter_range]{filter_range()}}).}

\item{add}{Should existing filter state specifications be added to?
Default is TRUE. If FALSE, the entire sort specification will be overridden.}
}
\description{
Add a filter state specifications to a trelliscope display
}
