% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_info_html}
\alias{set_info_html}
\title{Specify custom "display info" html}
\usage{
set_info_html(trdf, file)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{file}{Path to an existing html file to use.}
}
\description{
Specify custom "display info" html
}
