% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_labels}
\alias{set_labels}
\title{Add a labels state specification to a trelliscope display}
\usage{
set_labels(disp, varnames)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{trelliscope()}.}
}
\description{
Add a labels state specification to a trelliscope display
}
