% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_layout}
\alias{set_layout}
\title{Add a layout state specification to a trelliscope display}
\usage{
set_layout(disp, nrow = 1, ncol = 1, arrange = "rows", page = 1)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{trelliscope()}.}

\item{nrow}{Number of rows of panels to show.}

\item{ncol}{Number of columns of panels to show.}

\item{arrange}{Arrange panels by "rows" or "cols".}

\item{page}{The page number to show.}
}
\description{
Add a layout state specification to a trelliscope display
}
