% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_trelliscope.R
\name{set_panel_options}
\alias{set_panel_options}
\title{Set panel options for a Trelliscope display}
\usage{
set_panel_options(trdf, ...)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}}

\item{...}{A named set panel options to set. The names should
correspond to the names of the variables in the data frame. The values
should come from \code{\link[=panel_options]{panel_options()}}.}
}
\description{
Set panel options for a Trelliscope display
}
\examples{
mars_rover |>
  as_trelliscope_df(name = "mars rover") |>
  set_panel_options(img_src = panel_options(width = 2, height = 1))
}
