% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_primary_panel}
\alias{set_primary_panel}
\title{Set the primary panel of a trelliscope display}
\usage{
set_primary_panel(trdf, name)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{name}{The name of the panel to set as the primary panel that will be
viewed when launching the app.}
}
\description{
Set the primary panel of a trelliscope display
}
