% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_show_info_on_load}
\alias{set_show_info_on_load}
\title{Show "display info" when display first loads}
\usage{
set_show_info_on_load(trdf, show = TRUE)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{show}{Should display info be shown on load?}
}
\description{
Show "display info" when display first loads
}
