% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_sort}
\alias{set_sort}
\title{Add a labels state specification to a trelliscope display}
\usage{
set_sort(disp, varnames, dirs)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{trelliscope()}.}

\item{varnames}{A vector of variable names to sort on.}

\item{dirs}{A vector of directions to sort on ("asc" or "desc").}
}
\description{
Add a labels state specification to a trelliscope display
}
