% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_and_tags.R
\name{set_tags}
\alias{set_tags}
\title{Set tags for variables in a data frame}
\usage{
set_tags(d, ...)
}
\arguments{
\item{d}{A data frame.}

\item{...}{A named set of vectors of variable names, where each name is a
tag and each value in each vector corresponds to one of the variables in
the dataset.}
}
\description{
Set tags for variables in a data frame
}
\examples{
set_tags(iris,
  info = "Species",
  metrics = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
  width = c("Sepal.Width", "Petal.Width"))
}
