% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_theme}
\alias{set_theme}
\title{Set a color theme for a Trelliscope display}
\usage{
set_theme(
  trdf,
  primary = "#448aff",
  dark = "#2e60b1",
  light = "#4dabf5",
  light_text_on_dark = TRUE,
  dark_text = "#000000",
  light_text = "#ffffff",
  header_background = "#fefefe",
  header_text = NULL,
  logo = NULL
)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}}

\item{primary}{Primary color to use.}

\item{dark}{Dark color to use.}

\item{light}{Light color to use.}

\item{light_text_on_dark}{Should light text be used on backgrounds.
using the dark color?}

\item{dark_text}{Dark text color to use.}

\item{light_text}{Light text color to use.}

\item{header_background}{Color to use for the header background.}

\item{header_text}{Color to use for the header text.}

\item{logo}{URL (relative or absolute) to a logo image to
include in the header.}
}
\description{
Set a color theme for a Trelliscope display
}
