% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_and_tags.R
\name{set_var_labels}
\alias{set_var_labels}
\title{Set labels for variables in a data frame}
\usage{
set_var_labels(d, ...)
}
\arguments{
\item{d}{A data frame.}

\item{...}{A named set of labels, where each name must correspond to one
of the variables in the dataset.}
}
\description{
Set labels for variables in a data frame
}
\examples{
set_var_labels(mtcars, mpg = "Miles per gallon", cyl = "Number of cylinders")
}
