% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_wrappers.R
\name{state_labels}
\alias{state_labels}
\title{Specify a "labels" state}
\usage{
state_labels(varnames)
}
\arguments{
\item{labels}{A vector of variable names that should appear as labels
when viewing the display.}
}
\description{
Specify a "labels" state
}
