% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_wrappers.R
\name{state_layout}
\alias{state_layout}
\title{Specify a "layout" state}
\usage{
state_layout(ncol = 1, page = 1)
}
\arguments{
\item{ncol}{Number of columns of panels to show.}

\item{page}{The page number to show.}
}
\description{
Specify a "layout" state
}
