% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_wrappers.R
\name{state_sort}
\alias{state_sort}
\title{Specify a "sort" state}
\usage{
state_sort(varname, dir = "asc")
}
\arguments{
\item{varname}{The name of the variable to sort on.}

\item{dir}{One of "asc" or "desc", describing the direction of the sort.}
}
\description{
Specify a "sort" state
}
