% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qtrellis.R
\name{to_ddf}
\alias{to_ddf}
\title{Convert dplyr grouped_df to ddf}
\usage{
to_ddf(x)
}
\arguments{
\item{x}{a grouped_df object from dplyr}
}
\description{
Convert dplyr grouped_df to ddf
}
\examples{
\dontrun{
library(dplyr)
bySpecies <- iris \%>\%
  group_by(Species) \%>\%
  to_ddf()
}
}

