% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscope-package.R
\docType{package}
\name{trelliscope-package}
\alias{trelliscope}
\alias{trelliscope-package}
\title{trelliscope}
\description{
Trelliscope provides a way to create, organize, and view
multi-panel displays, a collection of which is called a visualization
database (VDB).  This package provides a mechanism for extending the ideas
of Trellis Display to easily generate and organize plots, providing a
dynamic viewer for tiling panels of a display across the screen and paging
through them. There is also a mechanism to generate built-in or
user-defined "cognostic" metrics for each panel of a display , and the
viewer has the ability to sort or filter the pages based on these
cognostics.  This is useful for large multi-panel displays where it is not
possible to view each panel, allowing only panels with interesting features
to be presented.
}
\details{
\url{http://tessera.io/docs-trelliscope/}
}
\examples{
help(package = trelliscope)
}
\author{
Ryan Hafen
Maintainer: Ryan Hafen <rhafen@gmail.com>
}
\keyword{package}

