% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{trelliscope-shiny}
\alias{trelliscope-shiny}
\alias{trelliscopeOutput}
\alias{renderTrelliscope}
\title{Shiny bindings for trelliscope}
\usage{
trelliscopeOutput(outputId, ...)

renderTrelliscope(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{...}{Argumentsed passed on to \code{\link[shiny:htmlOutput]{shiny::htmlOutput()}}
is useful if you want to save an expression in a variable.}

\item{expr}{An expression that generates a trelliscope}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This}
}
\description{
Output and render functions for using trelliscope within Shiny
applications and interactive Rmd documents.
}
