% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_write.R
\name{update_display_list}
\alias{update_display_list}
\title{Update the list of all displays in an app directory}
\usage{
update_display_list(app_path, jsonp = TRUE, id)
}
\arguments{
\item{app_path}{The path where all of the displays are stored}

\item{jsonp}{If true, files are read and written as "jsonp" format,
otherwise "json" format. The "jsonp" format makes it possible to browse a
trelliscope app without the need for a web server.}

\item{id}{The id of the display. Can be found in \code{config.json[p]}.}
}
\description{
Update the list of all displays in an app directory
}
