% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDisplay.R
\name{updateDisplay}
\alias{updateDisplay}
\title{Smart Update Display}
\usage{
updateDisplay(name, group = "common", conn = getOption("vdbConn"),
  mode = c("update", "make", "auto"), ...)
}
\arguments{
\item{name}{See makeDisplay}

\item{group}{See makeDisplay}

\item{conn}{See makeDisplay}

\item{mode}{if "update" just tries to update display, if "make" just tries to makeDisplay, and finally if "auto" tries to update and failing that tries to "make".}

\item{...}{parameters updated from makeDisplay}
}
\value{
Depends on updateOrMake
}
\description{
Can be used to automatically make or update display depending on if it exists (avoiding cognostic recalculation).
}

