% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{vdbCopyRSource}
\alias{vdbCopyRSource}
\title{Copy files ending in .R from source directory into a VDB-wide global code directory for use in Trelliscope displays}
\usage{
vdbCopyRSource(fromDir, conn = getOption("vdbConn"))
}
\arguments{
\item{fromDir}{directory with R source}

\item{conn}{vdb connection}
}
\description{
These R files will be sourced in the global environment of the Trelliscope Viewer when the viewer is launched and the resulting objects will be available to all displays throughout the viewing session.
Useful for getting custom functions into the global environment.
}
\examples{
\dontrun{
vdbCopyRSource(".") # copies R files from current directory
}

}
\author{
Jeremiah Rounds
}

