% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{vdbGlobalsExist}
\alias{vdbGlobalsExist}
\title{Check to see if the VDB-wide global data file exists}
\usage{
vdbGlobalsExist(conn = getOption("vdbConn"))
}
\arguments{
\item{conn}{vdb connection}
}
\value{
logical
}
\description{
Check to see if the VDB-wide global data file exists
}
\examples{
vdbGlobalsExist()
}
\author{
Jeremiah Rounds
}

