% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGlobals.R
\name{vdbGlobalsFile}
\alias{vdbGlobalsFile}
\title{Path to VDB global data storage file}
\usage{
vdbGlobalsFile(conn = getOption("vdbConn"))
}
\arguments{
\item{conn}{vdb connection}
}
\value{
character file name
}
\description{
Returns an appropriate file name to save VDB-wide globals to for use in Trelliscope displays
}
\details{
Objects in this rdata file will be loaded by the Trelliscope Viewer when the viewer is launched and the data will be available to all displays throughout the viewing session.
}
\examples{
\dontrun{
save(foo, file = globalsPath())
}
}
\author{
Jeremiah Rounds
}

