\name{vdbInit}
\alias{vdbInit}
\title{Set up a new VDB}
\usage{
  vdbInit(path = NULL, name = "myVDB", autoYes = FALSE)
}
\arguments{
  \item{path}{The location on the local file system where
  the directory for the vdb will be placed}

  \item{autoYes}{should questions to proceed with
  operations be automarically answered with "yes"?}
}
\value{
  nothing
}
\description{
  Set up a new vdb directory
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{updateViewer}}
}

