\name{vdbUpdateConn}
\alias{vdbUpdateConn}
\title{Update VDB Connection Options}
\usage{
  vdbUpdateConn(conn)
}
\arguments{
  \item{conn}{A list with names the correspond to VDB
  connection options and the corresponding settings}
}
\value{
  nothing
}
\description{
  Update VDB connection options
}
\details{
  It is easier to simply edit the conn.R file in your VDB
  directory
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{vdbConnect}}, \code{\link{vdbInit}}
}

