% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{view_display}
\alias{view_display}
\title{View a trelliscope display}
\usage{
view_display(disp)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{\link[=trelliscope]{trelliscope()}}.}
}
\description{
View a trelliscope display
}
