% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_view.R
\name{view_trelliscope}
\alias{view_trelliscope}
\title{View a trelliscope display}
\usage{
view_trelliscope(trdf = NULL)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}
}
\description{
View a trelliscope display
}
\examples{
library(ggplot2)

panel_dat <- (
  ggplot(gap, aes(year, life_exp)) +
    geom_point() +
    facet_panels(vars(country, continent))
) |>
  as_panels_df()

disp <- panel_dat |>
  as_trelliscope_df(name = "life_expectancy")

\dontrun{
view_trelliscope(disp)
}
}
