% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_write.R
\name{write_trelliscope}
\alias{write_trelliscope}
\title{Write the contents of a display}
\usage{
write_trelliscope(trdf, force_write = NULL, jsonp = NULL)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{force_write}{Should the panels be forced to be written even if they
have already been written?}

\item{jsonp}{If true, app files are written as "jsonp" format, otherwise
"json" format. The "jsonp" format makes it possible to browse a
trelliscope app without the need for a web server.}
}
\description{
Write the contents of a display
}
\examples{
library(ggplot2)

panel_dat <- (
  ggplot(gap, aes(year, life_exp)) +
    geom_point() +
    facet_panels(vars(country, continent))
) |>
  as_panels_df()

disp <- panel_dat |>
  as_trelliscope_df(name = "life_expectancy")

\dontrun{
disp <- write_trelliscope(disp)
view_trelliscope(disp)
}
}
