% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_charm}
\alias{add_charm}
\title{Use fidelius to password protect a trelliscope display}
\usage{
add_charm(trdf, ...)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope]{as_trelliscope()}} or a
data frame which will be cast as such.
"trelliscopejs_widget".}

\item{...}{Arguments passed to the charm() function in the fidelius package.}
}
\description{
Use fidelius to password protect a trelliscope display
}
