% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_input_email}
\alias{add_input_email}
\title{Specify an email address to which input feedback can be sent}
\usage{
add_input_email(trdf, email)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{email}{An email address.}
}
\description{
Specify an email address to which input feedback can be sent
}
