% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_input_vars}
\alias{add_input_vars}
\title{Specify meta variables whose values should be provided in input feedback}
\usage{
add_input_vars(trdf, vars)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{vars}{A vector of meta variable names found in the display.}
}
\description{
Specify meta variables whose values should be provided in input feedback
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

panel_dat <- (ggplot(gapminder, aes(year, lifeExp)) +
  geom_point() +
  facet_panels()) |>
  nest_panels()
  
trell <- panel_dat |>
as_trelliscope_df() |>
write_panels() |>
add_inputs(
  input_radio(name = "Radio Input", 
               label = "A space to add custom ranking for sorting",
               options = c("yes", "no"))) |>
add_input_vars(c("continent", "country")) |>
add_input_email("johndoe@email.com") |>
write_trelliscope() |>
view_trelliscope()
}


}
