% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_inputs}
\alias{add_inputs}
\title{Add inputs to a trelliscope display}
\usage{
add_inputs(trdf, ...)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope]{as_trelliscope()}} or a
data frame which will be cast as such.}

\item{...}{Any number of input specifications. These can be specified with
any of \code{\link[=input_number]{input_number()}}, \code{\link[=input_radio]{input_radio()}}, \code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_select]{input_select()}}, \code{\link[=input_multiselect]{input_multiselect()}}, \code{\link[=input_text]{input_text()}}}
}
\description{
Add inputs to a trelliscope display
}
