% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_meta_def}
\alias{add_meta_def}
\title{Add a meta variable definition to a trelliscope display}
\usage{
add_meta_def(disp, obj)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{trelliscope()}.}

\item{obj}{A meta variable definition created with a meta_*() function}
}
\description{
Add a meta variable definition to a trelliscope display
}
