% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{add_meta_tags}
\alias{add_meta_tags}
\title{Specify tags for meta variables}
\usage{
add_meta_tags(trdf, ...)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}} or a
data frame which will be cast as such.}

\item{...}{A named set of vectors of tags, where each name must correspond to
one of the variables in the dataset}
}
\description{
Specify tags for meta variables
}
\details{
This function can be useful if you don't want to go to the trouble
of explicitly setting meta variable definitions but still want variable
tags.
}
