% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applyCogFn}
\alias{applyCogFn}
\title{Apply Cognostics Function to a Key-Value Pair}
\usage{
applyCogFn(cogFn, kvSubset, conn)
}
\arguments{
\item{cogFn}{cognostics function}

\item{kvSubset}{key-value pair}

\item{conn}{TODO}
}
\description{
Apply cognostics function to a key-value pair
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{cog}}, \code{\link{makeDisplay}}
}

