% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{as_json}
\alias{as_json}
\title{Convert any trelliscope R6 object to JSON}
\usage{
as_json(df, pretty = TRUE)
}
\arguments{
\item{pretty}{Adds indentation whitespace to JSON output. Can be TRUE/FALSE
or a number specifying the number of spaces to indent.}

\item{obj}{Any R6 object created with trelliscope functions.}
}
\description{
Convert any trelliscope R6 object to JSON
}
