% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trelliscope-data.R
\docType{data}
\name{batting}
\alias{batting}
\title{Batting Data}
\format{A data frame with 1406 rows and 22 variables}
\usage{
data(batting)
}
\description{
Batting data
}
\details{
\itemize{
  \item playerID. : Factor w/ 17908 levels "aardsda01","aaronha01",..: 1 6 34 34 37 40 40 41 46 51 ...
  \item yearID.   : int  2012 2012 2012 2012 2012 2012 2012 2012 2012 2012 ...
  \item stint.    : int  1 1 1 2 1 1 2 1 1 1 ...
  \item teamID.   : Factor w/ 149 levels "ALT","ANA","ARI",..: 93 58 71 72 66 45 96 16 116 94 ...
  \item lgID.     : Factor w/ 6 levels "AA","AL","FL",..: 2 4 2 4 2 2 2 2 2 4 ...
  \item G.        : int  1 37 8 92 22 26 1 69 153 45 ...
  \item AB.       : int  NA 7 24 195 70 0 NA 0 607 0 ...
  \item R.        : int  NA 0 1 28 5 0 NA 0 84 0 ...
  \item H.        : int  NA 1 5 48 18 0 NA 0 137 0 ...
  \item X2B.      : int  NA 0 3 8 2 0 NA 0 22 0 ...
  \item X3B.      : int  NA 0 0 1 1 0 NA 0 2 0 ...
  \item HR.       : int  NA 0 0 3 1 0 NA 0 12 0 ...
  \item RBI.      : int  NA 0 5 19 15 0 NA 0 50 0 ...
  \item SB.       : int  NA 0 0 6 0 0 NA 0 13 0 ...
  \item CS.       : int  NA 0 0 2 0 0 NA 0 3 0 ...
  \item BB.       : int  NA 0 2 35 2 0 NA 0 59 0 ...
  \item SO.       : int  NA 3 5 51 13 0 NA 0 124 0 ...
  \item IBB.      : int  NA 0 0 3 0 0 NA 0 7 0 ...
  \item HBP.      : int  NA 0 0 0 1 0 NA 0 0 0 ...
  \item SH.       : int  NA 0 0 0 0 0 NA 0 1 0 ...
  \item SF.       : int  NA 0 1 0 1 0 NA 0 1 0 ...
  \item GIDP.     : int  NA 1 1 6 1 0 NA 0 3 0 ...
  \item price. price in US dollars (\$326--\$18,823)
}
}
\keyword{datasets}

