\name{cogNcol}
\alias{cogNames}
\alias{cogNcol}
\alias{cogNrow}
\alias{getCogData}
\alias{oldGetCurCogDat}
\title{Methods for Cognostics Connections}
\usage{
cogNcol(x, ...)

cogNrow(x, ...)

cogNames(x, ...)

getCogData(x, ...)

oldGetCurCogDat(x, ...)
}
\arguments{
  \item{x}{object}
}
\description{
Methods for Cognostics Connections
}
\note{
These methods are used mainly by the trelliscope viewer and
therefore must be exported.  They should never need to be
used by an analyst.
}

