% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_cog_methods.R
\name{cogNcol}
\alias{cogNames}
\alias{cogNcol}
\alias{cogNrow}
\alias{getCogData}
\title{Methods for Cognostics Connections}
\usage{
cogNcol(x, ...)

cogNrow(x, ...)

cogNames(x, ...)

getCogData(x, rowIdx, colIdx, ...)
}
\arguments{
\item{x}{object}

\item{...}{other objects passed onto generic methods}

\item{rowIdx}{TODO}

\item{colIdx}{TODO}
}
\description{
Methods for Cognostics Connections
}
\note{
These methods are used mainly by the trelliscope viewer and therefore must be exported.  They should never need to be used by an analyst.
}

