% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cogNcol}
\alias{cogNames}
\alias{cogNcol}
\alias{cogNrow}
\alias{getCogData}
\alias{oldGetCurCogDat}
\title{Methods for Cognostics Connections}
\usage{
cogNcol(x, ...)

cogNrow(x, ...)

cogNames(x, ...)

getCogData(x, rowIdx, colIdx, ...)

oldGetCurCogDat(cogDF, flt, ordering, colIndex, verbose, ...)
}
\arguments{
\item{x}{object}

\item{...}{other objects passed onto generic methods}

\item{rowIdx}{TODO}

\item{colIdx}{TODO}

\item{cogDF}{TODO}

\item{flt}{TODO}

\item{ordering}{TODO}

\item{colIndex}{TODO}

\item{verbose}{TODO}
}
\description{
Methods for Cognostics Connections
}
\note{
These methods are used mainly by the trelliscope viewer and therefore must be exported.  They should never need to be used by an analyst.
}

