% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cognostics.R
\name{cogDisplayHref}
\alias{cogDisplayHref}
\title{DisplayHref Cognostic}
\usage{
cogDisplayHref(displayName, displayGroup = "common", state = NULL,
  label = "link", desc = "display link", group = "common",
  target = "_blank", defLabel = FALSE, defActive = FALSE,
  filterable = FALSE)
}
\arguments{
\item{displayName}{the name of the display}

\item{displayGroup}{the group the display belongs to}

\item{state}{if specified, this tells the viewer the default parameter settings (such as layout, sorting, filtering, etc.) to use when the display is viewed (see \code{\link{validateState}} for details)}

\item{label}{label of the href}

\item{desc,group,defLabel,defActive,filterable}{arguments passed to \code{\link{cog}}}

\item{target}{value to be used for the \code{target} attribute of the \code{a} html tag - default is "_blank" which will open the link in a new window}
}
\value{
a hash string
}
\description{
Create href that points to another trelliscope display with optional state
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{validateState}}, \code{\link{cogHref}}
}

