% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cognostics.R
\name{cogHref}
\alias{cogHref}
\title{Href Cognostic}
\usage{
cogHref(x, label = "link", desc = "link", group = "common",
  target = "_blank", defLabel = FALSE, defActive = FALSE,
  filterable = FALSE, sortable = TRUE, log = FALSE)
}
\arguments{
\item{x}{URL to link to}

\item{label}{label of the href}

\item{desc,group,defLabel,defActive,filterable,sortable,log}{arguments passed to \code{\link{cog}}}

\item{target}{value to be used for the \code{target} attribute of the \code{a} html tag - default is "_blank" which will open the link in a new window}
}
\description{
Create href to be used as cognostics in a trelliscope display.
}
\examples{
cogHref("www.google.com")
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{cog}}
}

