% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{cogLoessRMSE}
\alias{cogLoessRMSE}
\title{Compute RMSE of Loess Fit Cognostic}
\usage{
cogLoessRMSE(..., desc = "RMSE of residuals from loess fit",
  group = "common", defLabel = FALSE, defActive = TRUE,
  filterable = TRUE, sortable = TRUE, log = FALSE)
}
\arguments{
\item{desc, group, defLabel, defActive, filterable, sortable, log}{arguments passed to \code{\link{cog}}}

\item{\ldots}{arguments to be passed to \code{link{loess}}, such as the formula, data, smoothing parameters, etc.}
}
\description{
Compute RMSE of loess fit as a cognostic to be used as cognostics in a trelliscope display.
}
\examples{
cogLoessRMSE(dist ~ speed, span = 0.5, data = cars)
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{cog}}
}

