% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cogLoessRMSE}
\alias{cogLoessRMSE}
\title{Compute RMSE of Loess Fit Cognostic}
\usage{
cogLoessRMSE(..., desc = "RMSE of residuals from loess fit",
  group = "common", defLabel = FALSE, defActive = TRUE,
  filterable = TRUE)
}
\arguments{
\item{\ldots}{arguments to be passed to \code{link{loess}}, such as the formula, data, smoothing parameters, etc.}

\item{desc,group,defLabel,defActive,filterable}{arguments passed to \code{\link{cog}}}
}
\description{
Compute RMSE of loess fit as a cognostic to be used as cognostics in a trelliscope display.
}
\examples{
cogLoessRMSE(dist ~ speed, span = 0.5, data = cars)
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{cog}}
}

