% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cogScagnostics}
\alias{cogScagnostics}
\title{Compute Scagnostics}
\usage{
cogScagnostics(x, y)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}
}
\description{
Compute list of scagnostics (see \code{\link[scagnostics]{scagnostics}}) to be used as cognostics in a trelliscope display.
}
\examples{
cogScagnostics(cars$dist, cars$speed)
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{cog}}
}

