% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cogScagnostics}
\alias{cogScagnostics}
\title{Compute Scagnostics}
\usage{
cogScagnostics(x, y, group = "scagnostics", defLabel = FALSE,
  defActive = TRUE, filterable = TRUE)
}
\arguments{
\item{x}{vector of the x-axis data for a scatterplot}

\item{y}{vector of the y-axis data for a scatterplot}

\item{group,defLabel,defActive,filterable}{arguments passed to \code{\link{cog}}}
}
\description{
Compute list of scagnostics (see \code{\link[scagnostics]{scagnostics}}) to be used as cognostics in a trelliscope display.
}
\examples{
cogScagnostics(iris$Sepal.Length, iris$Sepal.Width)
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{cog}}
}

