% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{toHash}
\alias{fromHash}
\alias{toHash}
\title{Methods for dealing with state and hashes}
\usage{
toHash(x)

fromHash(x)
}
\arguments{
\item{x}{state or hash object}
}
\description{
Methods for dealing with state and hashes
}

