% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer_meta}
\alias{infer_meta}
\title{Infer meta variable definitions}
\usage{
infer_meta(trdf)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}
}
\description{
Infer meta variable definitions
}
