% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_multiselect}
\alias{input_multiselect}
\title{Specify a "multiselect" input}
\usage{
input_multiselect(name, label = name, active = TRUE, options)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{options}{A vector of options for the multiselect dropdown.}
}
\description{
Specify a "multiselect" input
}
\examples{
library(ggplot2)
library(dplyr)

panel_dat <- (
  ggplot(gap, aes(year, life_exp)) +
    geom_point() +
    facet_panels(vars(country, continent))
  ) |>
  as_panels_df()

trell <- panel_dat |>
as_trelliscope_df() |>
add_inputs(
  input_multiselect(
    name = "Multiselect Input",
    label = "A space to add custom dropdown inputs",
    options =c("yes", "no")
  )
)
\dontrun{
view_trelliscope(trell)
}
}
\seealso{
Other inputtypes: 
\code{\link{input_checkbox}()},
\code{\link{input_number}()},
\code{\link{input_radio}()},
\code{\link{input_select}()},
\code{\link{input_text}()}
}
\concept{inputtypes}
