% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_number}
\alias{input_number}
\title{Specify a "numeric" input box}
\usage{
input_number(name, label = name, active = TRUE, min = NULL, max = NULL)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{min}{Optional minimum value to allow in the input.}

\item{max}{Optional maximum value to allow in the input.}
}
\description{
Specify a "numeric" input box
}
\seealso{
Other {input types}: 
\code{\link{input_checkbox}()},
\code{\link{input_multiselect}()},
\code{\link{input_radio}()},
\code{\link{input_select}()},
\code{\link{input_text}()}
}
\concept{{input types}}
