% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_text}
\alias{input_text}
\title{Specify a "text box" input}
\usage{
input_text(name, label = name, active = TRUE, width = 80, height = 3)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{width}{Width (in characters) of the text box input.}

\item{height}{Height (in lines of text) of the text box input.}
}
\description{
Specify a "text box" input
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

panel_dat <- (
  ggplot(gapminder, aes(year, lifeExp)) +
    geom_point() +
    facet_panels()
  ) |>
  nest_panels()

trell <- panel_dat |>
  as_trelliscope_df() |>
  write_panels() |>
  add_inputs(
    input_text(
      name = "Text Input",
      label = "A space to add custom text input")
    ),
    email = "johndoe@email.com"
)
}
}
\seealso{
Other {input types}: 
\code{\link{input_checkbox}()},
\code{\link{input_multiselect}()},
\code{\link{input_number}()},
\code{\link{input_radio}()},
\code{\link{input_select}()}
}
\concept{{input types}}
