% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrappers.R
\name{input_text}
\alias{input_text}
\title{Specify a "text box" input}
\usage{
input_text(name, label = name, active = TRUE, height = 3)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Description of the input.}

\item{active}{Should the input be active by default?}

\item{height}{Height (in lines of text) of the text box input.}
}
\description{
Specify a "text box" input
}
\examples{
library(ggplot2)
library(dplyr)

panel_dat <- (
  ggplot(gap, aes(year, life_exp)) +
    geom_point() +
    facet_panels(vars(country, continent))
  ) |>
  as_panels_df()

trell <- panel_dat |>
  as_trelliscope_df() |>
  add_inputs(
    input_text(
      name = "Text Input",
      label = "A space to add custom text input"
    )
)
\dontrun{
view_trelliscope(trell)
}
}
\seealso{
Other {input types}: 
\code{\link{input_checkbox}()},
\code{\link{input_multiselect}()},
\code{\link{input_number}()},
\code{\link{input_radio}()},
\code{\link{input_select}()}
}
\concept{{input types}}
