% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splod.R
\name{makeSplodData}
\alias{makeSplodData}
\title{Create Data Plottable by splod}
\usage{
makeSplodData(data, id.vars = NULL, ...)
}
\arguments{
\item{data}{a data.frame}

\item{id.vars}{variables to ignore when computing all pairs of variables}

\item{...}{TODO}
}
\value{
an object of class 'localDiv' and 'splodDat' that can be passed to \code{\link{splod}}
}
\description{
Create pairwise scatterplot data plottable by splod
}
\references{
Wilkinson, L., Anushka A., and Grossman, R. L. "Graph-Theoretic Scagnostics." INFOVIS. Vol. 5. 2005.
}
\seealso{
\code{\link{splod}}, \code{\link{splodPanelFn}}
}

