% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state.R
\name{makeStateHash}
\alias{makeStateHash}
\title{Make a URL hash out of state information}
\usage{
makeStateHash(x, name = NULL, group = NULL)
}
\arguments{
\item{x}{a list of cognostics state parameters}

\item{name}{the name of the display}

\item{group}{the group the display belongs to}
}
\value{
a URL hash
}
\description{
Make a URL hash out of cognostics state information
}
\examples{
x <- list(
  sort = list(state = "desc", county = "asc"),
  filter = list(
    county = list(regex = "Bent"),
    state = list(select = "WA"),
    meanList = list(from = 0, to = 1)
  ),
  layout = list(nrow = 2, ncol = 4),
  labels = c("county", "state")
)
makeStateHash(validateState(x))
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{validateState}}
}

