% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_geo}
\alias{meta_geo}
\title{Specify a "geo" metadata variable}
\usage{
meta_geo(varname, description = varname, tags = "common", latvar, longvar)
}
\arguments{
\item{varname}{Name of the variable.}

\item{description}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}

\item{latvar}{Name of variable that contains the latitude.}

\item{longvar}{Name of variable that contains the longitude.}
}
\description{
Specify a "geo" metadata variable
}
\seealso{
Other {metadata types}: 
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()},
\code{\link{meta_string}()}
}
\concept{{metadata types}}
