% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wrappers.R
\name{meta_string}
\alias{meta_string}
\title{Specify a "string" metadata variable}
\usage{
meta_string(varname, label = NULL, tags = NULL)
}
\arguments{
\item{varname}{Name of the variable.}

\item{label}{Description of the variable.}

\item{tags}{Vector of tag names that help classify this variable.}
}
\description{
Specify a "string" metadata variable
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

panel_dat <- (
  ggplot(gapminder, aes(year, lifeExp)) +
    geom_point() +
    facet_panels(~country + continent)
  ) |>
  nest_panels()

meta_dat <- gapminder|>
  group_by(country, continent) |>
  summarise(
    mean_lifeexp = mean(lifeExp),
    min_lifeexp = min(lifeExp),
    max_lifeexp = max(lifeExp),
    mean_gdp = mean(gdpPercap),
    first_year = min(year),
    latitude = first(latitude),
    longitude = first(longitude),
    .groups = "drop"
  ) |>
  ungroup() |>
  mutate(
    first_date = as.Date(paste0(first_year, "-01-01")),
    first_datetime = as.POSIXct.Date(first_date),
    continent = as.factor(continent),
    wiki_link = paste0("https://en.wikipedia.org/wiki/", country)
  )

joined_dat <- left_join(panel_dat, meta_dat,
  by = join_by(country, continent))

trell <- joined_dat |>
  as_trelliscope_df() |>
  write_panels() |>
  add_meta_defs(
    meta_string(
      "country",
      label = "Name of the Country",
      tags = "More Info"
    )
  )
}
}
\seealso{
Other {metadata types}: 
\code{\link{meta_currency}()},
\code{\link{meta_datetime}()},
\code{\link{meta_date}()},
\code{\link{meta_factor}()},
\code{\link{meta_geo}()},
\code{\link{meta_graph}()},
\code{\link{meta_href}()},
\code{\link{meta_number}()}
}
\concept{{metadata types}}
