% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cogPre}
\alias{cogCollect}
\alias{cogEmit}
\alias{cogFinal}
\alias{cogPre}
\title{Methods Used in MapReduce for makeDisplay}
\usage{
cogPre(obj, ...)

cogEmit(obj, ...)

cogCollect(obj, ...)

cogFinal(obj, ...)
}
\arguments{
\item{obj}{object}

\item{obj}{object}

\item{obj}{object}

\item{obj}{object}
}
\description{
Methods Used in MapReduce for makeDisplay

Methods Used in MapReduce for makeDisplay

Methods Used in MapReduce for makeDisplay

Methods Used in MapReduce for makeDisplay
}

