% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_cog_methods.R
\name{cogPre}
\alias{cogCollect}
\alias{cogEmit}
\alias{cogFinal}
\alias{cogPre}
\title{Methods Used in MapReduce for makeDisplay}
\usage{
cogPre(cogConn, conn, group, name, ...)

cogEmit(cogConn, ...)

cogCollect(cogConn, ...)

cogFinal(cogConn, ...)
}
\arguments{
\item{cogConn, conn, group, name, \ldots}{TODO}
}
\description{
Methods Used in MapReduce for makeDisplay

Methods Used in MapReduce for makeDisplay

Methods Used in MapReduce for makeDisplay

Methods Used in MapReduce for makeDisplay
}
\examples{
# used internally when calling makeDisplay
}

