% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_panels.R
\name{nest_panels}
\alias{nest_panels}
\title{Render the panels of a trelliscope display}
\usage{
nest_panels(
  x,
  data_col = "data",
  panel_col = "panel",
  unnest_cols = NULL,
  as_plotly = FALSE,
  plotly_args = NULL,
  plotly_cfg = NULL
)
}
\arguments{
\item{x}{A ggplot object created with \code{\link[=facet_panels]{facet_panels()}}.}

\item{data_col}{The name of the column to store the nested data in.}

\item{panel_col}{The name of the column to store the rendered panels in.}

\item{unnest_cols}{An optional vector of extra variable names in \code{x}
to not be nested in the data. If specified, cannot vary within the
each combination of the specified facet variables.}

\item{as_plotly}{should the panels be written as plotly objects?}

\item{plotly_args}{optional named list of arguments to send to \code{ggplotly}}

\item{plotly_cfg}{optional named list of arguments to send to plotly's
`config`` method.}
}
\description{
Render the panels of a trelliscope display
}
