% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_panels.R
\name{panel_lazy}
\alias{panel_lazy}
\title{Add a "calculated panel" column to a dataset}
\usage{
panel_lazy(plot_fn, data = dplyr::pick(dplyr::everything()))
}
\arguments{
\item{plot_fn}{A function that produces a panel from a given subset of
\code{data}, taking as arguments any variable names in \code{data} that you would
like to be available in the plottinf function.}

\item{data}{A data frame from which subsets will be extracted and plots will
be made. Should be a superset of the summary dataset to which this plot
column is being added.}
}
\description{
Add a "calculated panel" column to a dataset
}
