% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_trelliscope.R
\name{panel_options}
\alias{panel_options}
\title{Specify options for lazily-rendered panels in a Trelliscope display}
\usage{
panel_options(aspect = 1)
}
\arguments{
\item{aspect}{The aspect ratio of the panel. This is the height divided
by the width and specifies the relative dimension of the image in viewer.}
}
\description{
Specify options for lazily-rendered panels in a Trelliscope display
}
\examples{
mars_rover |>
  as_trelliscope_df(name = "mars rover") |>
  set_panel_options(img_src = panel_options(aspect = 0.5))
}
