% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayObj.R
\name{print.displayObj}
\alias{print.displayObj}
\title{Print a display object}
\usage{
\method{print}{displayObj}(x, ...)
}
\arguments{
\item{x}{an object of class "displayObj"}

\item{\ldots}{further arguments passed to or from other methods}
}
\description{
Print a display object
}

