% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeDisplay}
\alias{removeDisplay}
\title{Remove a Display from a VDB}
\usage{
removeDisplay(name, group, conn = getOption("vdbConn"))
}
\arguments{
\item{name}{the name of the display}

\item{group}{the group of the display}

\item{conn}{VDB connection info, typically stored in options("vdbConn") at the beginning of a session, and not necessary to specify here if a valid "vdbConn" object exists}

\item{verbose}{logical - print messages about what is being done}
}
\description{
Remove a display from a VDB.
}
\details{
If a display is uniquely determined by its name, then group is not required.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{makeDisplay}}, \code{\link{removeDisplay}}
}

