% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/displayObj.R
\name{removeDisplay}
\alias{removeDisplay}
\title{Remove a Display from a VDB}
\usage{
removeDisplay(name = NULL, group = NULL, conn = getOption("vdbConn"),
  autoYes = FALSE, verbose = TRUE)
}
\arguments{
\item{name}{the name of the display}

\item{group}{the group of the display}

\item{conn}{VDB connection info, typically stored in options("vdbConn") at the beginning of a session, and not necessary to specify here if a valid "vdbConn" object exists}

\item{autoYes}{should questions to proceed with display removal be automatically answered with "yes"?}

\item{verbose}{logical - print messages about what is being done}
}
\description{
Remove a display from a VDB.
}
\details{
If a display is uniquely determined by its name, then group is not required.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{makeDisplay}}
}

