% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{set_default_layout}
\alias{set_default_layout}
\title{Add a layout state specification to a trelliscope display}
\usage{
set_default_layout(trdf, ncol = 1, page = 1)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}

\item{ncol}{Number of columns of panels to show.}

\item{page}{The page number to show.}
}
\description{
Add a layout state specification to a trelliscope display
}
