% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{show_info}
\alias{show_info}
\title{View trelliscope info of a trelliscope data frame}
\usage{
show_info(trdf)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}
}
\description{
View trelliscope info of a trelliscope data frame
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

panel_dat <- (ggplot(gapminder, aes(year, lifeExp)) +
                geom_point() +
                facet_panels(~country + continent)) |>
  nest_panels()

trell <- panel_dat |>
  as_trelliscope_df(name = "life expectancy", path = "gapminder") |>
  set_default_layout(nrow = 2, ncol = 4) |>
  write_panels() |>
  write_trelliscope()
  
show_info(trell)
  
}

}
