% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{show_info}
\alias{show_info}
\title{View trelliscope info of a trelliscope data frame}
\usage{
show_info(trdf)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}} or a
data frame which will be cast as such.}
}
\description{
View trelliscope info of a trelliscope data frame
}
