\name{splodCogFn}
\alias{splodCogFn}
\title{Default Cognostics Function for vdbSplod}
\usage{
  splodCogFn(df)
}
\arguments{
  \item{df}{a subset of data created by
  \code{\link{makeSplodData}}}
}
\value{
  a data.frame of scagnostics for the given subset
}
\description{
  Default cognostics function for vdbSplod
}
\author{
  Ryan Hafen
}
\references{
  Wilkinson's scagnostics paper.
}
\seealso{
  \code{\link{splodPlotFn}}, \code{\link{vdbSplod}},
  \code{\link{makeSplodData}}
}

