% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splod.R
\name{splodCogFn}
\alias{splodCogFn}
\title{Default Cognostics Function for splod}
\usage{
splodCogFn(df)
}
\arguments{
\item{df}{a subset of data created by \code{\link{makeSplodData}}}
}
\value{
a data.frame of scagnostics for the given subset
}
\description{
Default cognostics function for splod
}
\author{
Ryan Hafen
}
\references{
Wilkinson's scagnostics paper.
}
\seealso{
\code{\link{splodPanelFn}}, \code{\link{splod}}, \code{\link{makeSplodData}}
}

