% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splod.R
\name{splodPanelFn}
\alias{splodPanelFn}
\title{Default Plot Function for splod}
\usage{
splodPanelFn(df)
}
\arguments{
\item{df}{a subset of data created by \code{\link{makeSplodData}}}
}
\value{
a trellis plot object of a scatterplot for the given subset
}
\description{
Default plot function for splod
}
\references{
Wilkinson, L., Anushka A., and Grossman, R. L. "Graph-Theoretic Scagnostics." INFOVIS. Vol. 5. 2005.
}
\seealso{
\code{\link{splodCogFn}}, \code{\link{splod}}, \code{\link{makeSplodData}}
}

