\name{splodPlotFn}
\alias{splodPlotFn}
\title{Default Plot Function for splod}
\usage{
  splodPlotFn(df)
}
\arguments{
  \item{df}{a subset of data created by
  \code{\link{makeSplodData}}}
}
\value{
  a trellis plot object of a scatterplot for the given
  subset
}
\description{
  Default plot function for splod
}
\author{
  Ryan Hafen
}
\references{
  Wilkinson's scagnostics paper.
}
\seealso{
  \code{\link{splodCogFn}}, \code{\link{splod}},
  \code{\link{makeSplodData}}
}

