% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_wrappers.R
\name{state_labels}
\alias{state_labels}
\title{Specify a "labels" state}
\usage{
state_labels(varnames = NULL)
}
\arguments{
\item{varnames}{A vector of variable names whose values should appear
as labels. If NULL, no labels will be displayed.
when viewing the display.}
}
\description{
Specify a "labels" state
}
