% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscope.R
\name{trelliscope}
\alias{trelliscope}
\title{Instantiate a trelliscope display object}
\usage{
trelliscope(
  df,
  name,
  description = name,
  id_vars = NULL,
  path = tempfile(),
  force_plot = FALSE
)
}
\arguments{
\item{df}{A data frame that contains the metadata of the display as well as
a column that indicate the panels to be displayed.}

\item{name}{Name of the trelliscope display.}

\item{description}{Description of the trelliscope display.}

\item{id_vars}{Variable names in \code{df} that uniquely define a row of the
data. If not supplied, an attempt will be made to infer them.}

\item{path}{Directory in which to place the trelliscope display when
it is written using \code{\link[=write_display]{write_display()}}.}

\item{force_plot}{Should the panels be forced to be plotted, even if they
have already been plotted and have not changed since the previous plotting?}
}
\description{
Instantiate a trelliscope display object
}
