% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayObj.R
\name{updateDisplay}
\alias{updateDisplay}
\title{Update a Display Object}
\usage{
updateDisplay(name, ..., group = "common", conn = getOption("vdbConn"))
}
\arguments{
\item{name}{the name of the display}

\item{group}{the group the display belongs to}

\item{conn}{VDB connection info, typically stored in options("vdbConn") at the beginning of a session, and not necessary to specify here if a valid "vdbConn" object exists}

\item{\ldots}{display parameters to update which must be one of "desc", "width", "height", "keySig", "panelFn", "state" - see \code{\link{makeDisplay}} for details on these parameters.}
}
\description{
Update a Display Object
}

