% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View a Display or Run Shiny Display Viewer}
\usage{
view(name = NULL, group = NULL, state = NULL, openBrowser = TRUE,
  conn = getOption("vdbConn"), port = 8100L)
}
\arguments{
\item{name, group}{optional parameters to load the viewer with a pre-specified display - if not specified, the viewer will launch with a list to choose from}

\item{state}{an optional list of state variables to set the default viewing state for layout, sorting, filtering, and labels (see details)}

\item{openBrowser}{should the browser be automatically launched?}

\item{conn}{VDB connection info, typically stored in options("vdbConn") at the beginning of a session, and not necessary to specify here if a valid "vdbConn" object exists}

\item{port}{what port to use for the viewer}
}
\description{
View a display or run Shiny display viewer
}
\details{
The 'state'
}
\author{
Ryan Hafen
}

