% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{view_trelliscope}
\alias{view_trelliscope}
\title{View a trelliscope display}
\usage{
view_trelliscope(trdf = NULL)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope]{as_trelliscope()}} or a
data frame which will be cast as such.}
}
\description{
View a trelliscope display
}
