% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_view.R
\name{view_trelliscope}
\alias{view_trelliscope}
\title{View a trelliscope display}
\usage{
view_trelliscope(trdf = NULL)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope_df]{as_trelliscope_df()}}
or a data frame which will be cast as such.}
}
\description{
View a trelliscope display
}
\examples{
\dontrun{
library(ggplot2)

panel_dat <- (
  ggplot(gapminder, aes(year, lifeExp)) +
    geom_point() +
    facet_panels(~country + continent)
) |>
  nest_panels()

disp <- panel_dat |>
  as_trelliscope_df(name = "life_expectancy") |>
  write_panels() |>
  write_trelliscope() |>
  view_trelliscope()

# Alternatively you can build your trelliscope and call `view_trelliscope()`
# separately. This allows for fine tuning of the trelliscope without having
# to reopen it every time you make an edit.
trell <- (
  ggplot(gapminder, aes(year, lifeExp)) +
    geom_point() +
    facet_panels(~country + continent)
) |>
  nest_panels() |>
  as_trelliscope_df(name = "life_expectancy") |>
  write_panels() |>
  write_trelliscope()

view_trelliscope(trell)
}
}
