% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/webSync.R
\name{webSync}
\alias{webSync}
\title{Sync VDB and notebook Files to a Web Server}
\usage{
webSync(vdbConn = getOption("vdbConn"), webConn = getOption("vdbWebConn"),
  fixPermissions = FALSE, verbose = FALSE, rsync = NULL)
}
\arguments{
\item{vdbConn}{VDB connection settings}

\item{webConn}{web connection settings}

\item{fixPermissions}{should an attempt be made to fix permissions in the web directory?}

\item{verbose}{show rsync output}

\item{rsync}{location of rsync binary}
}
\description{
Sync VDB and notebook files to a web server
}
\details{
This requires rsync to be installed on your machine.  If you are syncing via ssh, this only works if public key authentication is enabled between your local machine and the remote server.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{typeset}}, \code{\link{webConn}}, \code{\link{syncLocalData}}
}

