% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_display.R
\name{write_display}
\alias{write_display}
\title{Write a trelliscope display object to disk.}
\usage{
write_display(disp)
}
\arguments{
\item{disp}{A trelliscope display object created with \code{trelliscope()}.}
}
\description{
Write a trelliscope display object to disk.
}
