% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_write.R
\name{write_trelliscope}
\alias{write_trelliscope}
\title{Write the contents of a display}
\usage{
write_trelliscope(trdf, force_write = FALSE, jsonp = TRUE)
}
\arguments{
\item{trdf}{A trelliscope data frame created with \code{\link[=as_trelliscope]{as_trelliscope()}} or a
data frame which will be cast as such.}

\item{force_write}{Should the panels be forced to be written even if they
have already been written?}

\item{jsonp}{If true, app files are written as "jsonp" format, otherwise
"json" format. The "jsonp" format makes it possible to browse a
trelliscope app without the need for a web server.}
}
\description{
Write the contents of a display
}
